// lib/api/schemas/educationLevel.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */

export type EducationLevel = typeof EducationLevel[keyof typeof EducationLevel];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EducationLevel = {
  SELF_TAUGHT: 'SELF_TAUGHT',
  BOOTCAMP: 'BOOTCAMP',
  HIGH_SCHOOL: 'HIGH_SCHOOL',
  UNDERGRAD: 'UNDERGRAD',
  MASTERS: 'MASTERS',
  PHD: 'PHD',
  OTHER: 'OTHER',
} as const;