// lib/api/schemas/sourceType.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */

export type SourceType = typeof SourceType[keyof typeof SourceType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SourceType = {
  BOOK: 'BOOK',
  PAPER: 'PAPER',
  VIDEO: 'VIDEO',
  WEB: 'WEB',
} as const;