// lib/api/schemas/studyPlanMode.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * Textbooked Backend API
 * Backend MVP API for textbook study planning and grading.
 * OpenAPI spec version: 1.0.0
 */

export type StudyPlanMode = typeof StudyPlanMode[keyof typeof StudyPlanMode];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StudyPlanMode = {
  GUIDED: 'GUIDED',
  SOURCE_FOCUSED: 'SOURCE_FOCUSED',
} as const;